refactor : 메서드명 변경 & 코드 컨벤션 적용
