--- src/main/java/chess/dto/InitializeDTO.java
@@ -1,28 +1,0 @@
-package chess.dto;
-
-public final class InitializeDTO {
-    private String roomId;
-    private String winner;
-    private String loser;
-
-    public InitializeDTO() {
-    }
-
-    public InitializeDTO(final String roomId, final String winner, final String loser) {
-        this.roomId = roomId;
-        this.winner = winner;
-        this.loser = loser;
-    }
-
-    public String getRoomId() {
-        return roomId;
-    }
-
-    public String getWinner() {
-        return winner;
-    }
-
-    public String getLoser() {
-        return loser;
-    }
-}
