+++ src/main/java/chess/service/ResultService.java
@@ -2,7 +2,7 @@ package chess.service;
 
 import chess.dao.ResultDAO;
 import chess.dao.UserDAO;
-import chess.dto.ResultDTO;
+import chess.dto.RankingDTO;
 import chess.dto.UserDTO;
 import org.springframework.stereotype.Service;
 
@@ -20,14 +20,14 @@ public final class ResultService {
         this.userDAO = userDAO;
     }
 
-    public List<ResultDTO> allUserResult() {
-        List<ResultDTO> results = new ArrayList<>();
+    public List<RankingDTO> allUserResult() {
+        List<RankingDTO> results = new ArrayList<>();
         List<UserDTO> users = userDAO.findAll();
         for (UserDTO user : users) {
             int userId = user.getId();
             int winCount = resultDAO.winCountByUserId(userId);
             int loseCount = resultDAO.loseCountByUserId(userId);
-            results.add(new ResultDTO(userDAO.findNicknameById(userId), winCount, loseCount));
+            results.add(new RankingDTO(userDAO.findNicknameById(userId), winCount, loseCount));
         }
         Collections.sort(results);
         return results;
